+++ Source/Engine/RHI/Vulkan/VulkanTypes/RHICommandBuffer.cpp
@@ -328,7 +328,7 @@ void RHICommandBuffer::CopyBuffer(const RHIBuffer& sourceBuffer, Uint64 sourceOf
 	vkCmdCopyBuffer2(m_cmdBufferHandle, &copyInfo);
 }
 
-void RHICommandBuffer::FillBuffer(const RHIBuffer& buffer, Uint64 offset, Uint64 range, Byte data)
+void RHICommandBuffer::FillBuffer(const RHIBuffer& buffer, Uint64 offset, Uint64 range, Uint32 data)
 {
 	SPT_PROFILER_FUNCTION();
 
@@ -336,7 +336,7 @@ void RHICommandBuffer::FillBuffer(const RHIBuffer& buffer, Uint64 offset, Uint64
 	SPT_CHECK(buffer.IsValid());
 	SPT_CHECK(offset + range <= buffer.GetSize());
 
-	vkCmdFillBuffer(m_cmdBufferHandle, buffer.GetBufferHandle(), offset, range, static_cast<Uint32>(data));
+	vkCmdFillBuffer(m_cmdBufferHandle, buffer.GetBufferHandle(), offset, range, data);
 }
 
 void RHICommandBuffer::BeginDebugRegion(const lib::HashedString& name, const lib::Color& color)
